+++ wls-gui-wahllokalsystem/src/App.vue
@@ -165,7 +165,6 @@ import {
   VTooltip,
 } from "vuetify/components";
 
-import { getUser } from "@/api/user-client";
 import TheBroadcastReadConfirmationDialog from "@/components/broadcast/TheBroadcastReadConfirmationDialog.vue";
 import WlsHeartbeat from "@/components/wlsComponents/WlsHeartbeat.vue";
 import { useBroadcastCronjobService } from "@/composables/broadcast/broadcastCronjobService.ts";
@@ -178,20 +177,12 @@ import {
   ROUTE_WAHLVORSTAND,
   TOAST,
 } from "@/constants";
-import { useUserStore } from "@/stores/userStore.ts";
-import { useWahlvorstandStore } from "@/stores/wahlvorstandStore";
-import { User, UserLocalDevelopment } from "@/types/User";
 
-const userStore = useUserStore();
-const wahlvorstandStore = useWahlvorstandStore();
-const { startBroadcastMessageInterval, stopBroadcastMessageInterval } =
-  useBroadcastCronjobService();
 const [drawer, toggleDrawer] = useToggle();
 const isOffline = ref(false);
+const { stopBroadcastMessageInterval } = useBroadcastCronjobService();
 
 onMounted(() => {
-  loadUser();
-
   // config for service worker indexed db (same config as in wahl-worker.js !)
   localforage.config({
     driver: localforage.INDEXEDDB,
@@ -205,26 +196,6 @@ onMounted(() => {
 onUnmounted(() => {
   stopBroadcastMessageInterval();
 });
-
-/**
- * Loads UserInfo from the backend and sets it in the store.
- */
-function loadUser(): void {
-  getUser()
-    .then((user: User) => userStore.setUser(user))
-    .catch(() => {
-      // No user info received, so fallback
-      if (import.meta.env.DEV) {
-        userStore.setUser(UserLocalDevelopment());
-      } else {
-        userStore.setUser(null);
-      }
-    })
-    .then(() => {
-      wahlvorstandStore.loadWahlvorstand();
-      startBroadcastMessageInterval();
-    });
-}
 </script>
 
 <style>
